bootstrap: install test deps only if the respective file exists
Some dracut RHEL branches have a dedicated file with additional
RHEL-specific dependencies needed for testing. Attempt to install
such dependencies only if the file exists instead of failing
(like with the current RHEL8 branch).

+++ slave/bootstrap.sh
@@ -88,7 +88,10 @@ esac
 
 if [[ $branch =~ RHEL-* ]]; then
     yum -y install https://dl.fedoraproject.org/pub/epel/epel-release-latest-7.noarch.rpm
-    yum -y install $(<test/test-rpms.txt)
+
+    if [[ -f test/test-rpms.txt ]]; then
+        yum -y install $(<test/test-rpms.txt)
+    fi
 fi
